Stateless components and HMR in mobx
Added a section to the best practices section of the docs covering stateless components and hot module reloading.

Gives an examples of dealing with stateless components and MobX in general and also provides a best practice for getting around the weird HMR quirks when working with a bunch of stateless components.