Fix examples so they can be copy-pasted
It isn't obvious for beginners to use the guide and not know what the imports should look like. 

Also, the `<RenderCounter />` component was missing from the example making copy-pasting throw and error. For people new to React/MobX, they may assume that component is builtin somehow so putting in the example makes it more explicit.