fix typescript tests
The `tape` package depends on electron, which depends on @types/node.
This caused a resolution conflict between node's declaration of
`require()` and the one in `test/typescript/require.d.ts`.

Also, since @types/node has a different return type signature for
setInterval than the one in the `dom` lib, we need to add a `as any`
cast to force the correct overload is resolved.