add an es-modules build (referenced with pkg.module)
this output file (mobx.module.js) is built with the ES2015 typescript target,
so no ES2015 features get transpiled, and es module exports are
preserved. this allows consumers of mobx to take advantage of
tree-shaking bundlers such as rollupjs and webpack 2.

I was considering using the `.mjs` extension, as that looks like it will
be the most likely extension for es modules, but as it's not supported
in node at all at the moment, I decided against it