fix access to global object (fixes #1109)
webpack UMD builds pass in the global object (`global` in node and
`window` in the browser) as a parameter named `global`. In RollupJS this
is not the case. Access to the global object has to be handled manually,
so let' just check for `window`, and if it doesn't exist fall back to
`global`.