Split IObservableFactories type definition and implementation
It seems the only reason to keep IObservableFactories as a class is to avoid declaring type definition and implementation separately. However, it does not seem to be entirely correct approach at runtime, because it's never expected that a new instance of IObservableFactories will be created. In addition, reading properties of prototype cause issues with some browsers.