Add memoization and getter access documentation (#2329)
* Add memoization and getter access documentation

This PR adds some paragraphs around computed caching and memoization that seem to be a stumbling block for new users of Mobx. I have been using Mobx for a few years coming from other derived data layers and this was never explicitly clear to me.  The use of the `get` keyword in example code with intention of not using the 

I had to ask for clarification about this on stackoverflow: https://stackoverflow.com/questions/61129530/mobx-how-to-cache-computed-values

and then pieced together tidbits from https://github.com/mobxjs/mobx/issues/356

* grammar

* expand on computed examples

* add keepalive section

* add details on keepalives

* autorun vs keepalive

* Fix computed total typo