Went through the docs to remove references to decorators.
Various external references are still out of date. Since the egghead
tutorials will be out of date, I've removed them for now, though
we could reinstate them with a note about mobx-decorators.

Since actions will be the new default, I had to reorganize the order
of things a bit to introduce them earlier on.

I've moved the flow typing support section to the FAQ and added a
note about TypeScript as well.