Clarify when reactions effect functions trigger.
The docs are imo not clear enough on when effect functions trigger. 

It's only after I read the notes on the `equals` option that I got  to know that there is an 'extra' check (it being the result of the data function) that determines if an effect is triggered.  I thought after reading the docs that the effect function would always be triggered if any of the dependencies change.

This makes it very easy to create all sorts of bugs if used incorrectly:
https://codesandbox.io/s/mobx-reaction-bug-tyssu?file=/src/index.ts

I added a "By default, the result of the _data_ function has to change in order for the _effect_ function to be triggered." in order to make the behavior more clear.