add comment to help react users avoid react bug
There's a pretty insidious bug where after following the original instructions, the app works seemingly fine. But it's not until the cache of the webpack server is reset that you find out that everything breaks. The error is not that specfiic and trackign it down to this babel change takes a while. Hopefully this will help others out there with the same problem. 

The location / wording could use work but just wanted to get this process started