Clear up documentation
We ran into a problem while installing mobx. We copied and used the original snippet and it broke our app (couldn't log or get hot-reload on react native). We got a warning about the 'loose': false but following this documentation didn't get us to where we needed to be.

I believe this is a clearer and more complete explanation of what is actually needed for configuration.